apps/_core/master/pyproject.toml
[tool.poetry]
name = "apps-core-master"
version = "0.1.0"
description = "Core master service for managing applications and services"
authors = ["MasterGowen <mastergowen@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
setuptools = "^69.0.0"
fastapi = ">=0.108.0,<1.0.0"
uvicorn = {version = ">=0.22.0,<0.23.0", extras = ["standard"]}
nicegui = "~=1.4.8"
python-keycloak = "~=3.1.0"
sqlalchemy = "~=2.0.23"
aiodocker = "~=0.21.0"
docker = "==6.1.3"
aiofiles = "~=23.2.1"
aiohttp = "~=3.9.1"
pyyaml = "~=6.0.1"
jinja2 = "~=3.1.2"
watchfiles = "~=0.21.0"
watchdog = "~=3.0.0"
croniter = "~=2.0.1"
python-multipart = "~=0.0.6"
bcrypt = "~=4.0.1"
requests = "<2.32.0"

pydantic-settings = "^2.12.0"
[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"
httpx = ">=0.24.0"
pytest-cov = ">=4.0.0"
coverage-badge = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "pytest:main"
test-cov = "pytest --cov=app:main"